% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{comfort}
\alias{comfort}
\title{Say consoling things}
\usage{
comfort()
}
\description{
Commiserate the failing of module tests.
}
\details{
(Idea borrowed from \code{devtools}.)
}
\seealso{
Other praise: \code{\link{celebrate}}
}
\concept{praise}
