% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{description_get}
\alias{description_get}
\title{Return DESCRIPTION}
\usage{
description_get(flpth)
}
\arguments{
\item{flpth}{Path to module}
}
\value{
named list
}
\description{
Return contents of DESCRIPTION
}
\seealso{
Other utils: \code{\link{examples_get}},
  \code{\link{file_create}}, \code{\link{pkgdetails_get}},
  \code{\link{pkgnm_get}}, \code{\link{string_replace}},
  \code{\link{tags_get}}, \code{\link{templates_get}},
  \code{\link{yaml_get}}
}
\concept{utils}
