% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_build}
\alias{docker_build}
\title{Build a Docker image}
\usage{
docker_build(img, tag, url_or_path, verbose)
}
\arguments{
\item{img}{Image name}

\item{tag}{Tag, e.g. 'latest'}

\item{url_or_path}{URL or file path to Dockerfile}

\item{verbose}{Be verbose? T/F}
}
\value{
Logical
}
\description{
Build a Docker image through a system call. Returns TRUE if
no errors are raised.
}
\examples{
library(outsider.devtools)

# # NOT RUN
# 
# # simplest possible Dockerfile
# df_text <- "
# FROM ubuntu:latest
# "
# 
# # create dir to host Dockerfile
# flpth <- file.path(tempdir(), 'test_docker_build')
# if(!dir.exists(flpth)) {
#   dir.create(flpth)
# }
# 
# # write to file
# write(x = df_text, file = file.path(flpth, 'Dockerfile'))
# 
# # run docker_build from flpth
# docker_build(img = 'test_docker_build', tag = 'latest', verbose = TRUE,
#              url_or_path = flpth)
}
\seealso{
Other docker: 
\code{\link{docker_push}()}
}
\concept{docker}
