% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_cmd}
\alias{docker_cmd}
\title{Send commands to Docker}
\usage{
docker_cmd(args, std_out = TRUE, std_err = TRUE)
}
\arguments{
\item{args}{Vector of arguments for "docker" command. E.g. '--help'.}

\item{std_out}{Logical or file path}

\item{std_err}{Logical or file path}
}
\value{
Logical
}
\description{
Safely send commands to Docker. All commands are also echo'd.
}
\seealso{
Other docker-private: 
\code{\link{docker_login}()}
}
\concept{docker-private}
