% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_login}
\alias{docker_login}
\title{Login to Docker}
\usage{
docker_login(username)
}
\arguments{
\item{username}{Username for Docker-Hub.}
}
\value{
Logical
}
\description{
Login to docker using username. User is prompted to provide
password.
Returns TRUE if no errors are raised.
}
\seealso{
Other docker-private: 
\code{\link{docker_cmd}()}
}
\concept{docker-private}
