% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_push}
\alias{docker_push}
\title{Push a Docker image to Docker Hub}
\usage{
docker_push(username, img, tag, verbose)
}
\arguments{
\item{username}{Login username for Docker Hub.}

\item{img}{Image name}

\item{tag}{Tag, e.g. 'latest'}

\item{verbose}{Be verbose? T/F}
}
\value{
Logical
}
\description{
Push a Docker image to Docker Hub. Requires a user to have login
details for Docker Hub, \url{https://hub.docker.com/}.
}
\seealso{
Other docker: 
\code{\link{docker_build}()}
}
\concept{docker}
