% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{examples_get}
\alias{examples_get}
\title{Determine examples}
\usage{
examples_get(flpth)
}
\arguments{
\item{flpth}{Path to module}
}
\value{
character
}
\description{
Determine example files for a module
}
\seealso{
Other utils: 
\code{\link{description_get}()},
\code{\link{file_create}()},
\code{\link{pkgdetails_get}()},
\code{\link{pkgnm_get}()},
\code{\link{string_replace}()},
\code{\link{tags_get}()},
\code{\link{templates_get}()},
\code{\link{yaml_get}()}
}
\concept{utils}
