% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{examples_test}
\alias{examples_test}
\title{Run each example of an outsider module}
\usage{
examples_test(flpth)
}
\arguments{
\item{flpth}{File path to location of module}
}
\value{
logical
}
\description{
Return TRUE if all of the outsider module functions successfully
run.
}
\seealso{
Other testing: \code{\link{test}}
}
\concept{testing}
