% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_create}
\alias{file_create}
\title{Create file}
\usage{
file_create(x, flpth)
}
\arguments{
\item{x}{Text for writing to file}

\item{flpth}{File path to be created}
}
\description{
Write x to a filepath. Forces creation of directories.
}
\seealso{
Other utils: \code{\link{description_get}},
  \code{\link{examples_get}}, \code{\link{pkgdetails_get}},
  \code{\link{pkgnm_get}}, \code{\link{string_replace}},
  \code{\link{tags_get}}, \code{\link{templates_get}},
  \code{\link{yaml_get}}
}
\concept{utils}
