% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fnames_get}
\alias{fnames_get}
\title{Function names for module}
\usage{
fnames_get(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
character vector
}
\description{
Return function names of all available functions for an
installed outsider modules
}
\seealso{
Other private: \code{\link{examples_test}},
  \code{\link{file_create}}, \code{\link{pkgdetails_get}},
  \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{test}}
}
\concept{private}
