% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_upload}
\alias{git_upload}
\title{Upload module to code-sharing service}
\usage{
git_upload(flpth, username, service = c("github", "gitlab", "bitbucket"))
}
\arguments{
\item{flpth}{File path to module.}

\item{username}{Username for code-sharing service.}

\item{service}{Code-sharing service}
}
\value{
Logical
}
\description{
Upload module to a git-based code-sharing service. Initiate
a git repo, add core module files, commit and push to remote.
}
\details{
Remote URL is determined to be: code sharing URL + username + R
package name.
Git must be configured on a user's system before this function can be run.
}
\examples{
# Git must be configured before function can be run.
# To configure, provide your name and email via the command-line, e.g.
# git config --global user.name "John Doe"
# git config --global user.email johndoe@example.com
# To check your username and email, try:
# git config --list

# NOT RUN
# # construct a simple module
# module_path <- module_skeleton(program_name = 'echo', flpth = getwd())
# module_check(flpth = module_path)
# module_identities(flpth = module_path)
# # after these steps all files are built, to upload to git use:
# git_upload(flpth = module_path, username = '[YOUR USERNAME]',
#            service = '[SERVICE]')
}
\seealso{
Other git: 
\code{\link{remote_git_exists}()}
}
\concept{git}
