% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_check}
\alias{module_check}
\title{Check names and structure of a module}
\usage{
module_check(flpth)
}
\arguments{
\item{flpth}{File path to location of module}
}
\value{
Logical
}
\description{
Returns TRUE if all the names and structure of a minimal viable
outsider module are correct.
}
\examples{
library(outsider)
# build a skeleton package
flpth <- tempdir()
mdl_pth <- module_skeleton(program_name = 'example', repo_user = 'github_user',
                           docker_user = 'docker_user', flpth = flpth)
# look-up module user and program name/ids
# note: "om.." is added to the beginning of the program name
# mdl_pth <- file.path(flpth, 'om..example')
# (ids <- module_identities(flpth = mdl_pth))
# # generate .travis.yml
# module_travis(repo = ids$`GitHub repo`, flpth = mdl_pth)
# cat(readLines(file.path(mdl_pth, '.travis.yml')), sep = '\\n')
# # check the package is formatted correctly
# module_check(flpth = mdl_pth)


# clean-up
unlink(x = mdl_pth, recursive = TRUE)
}
\seealso{
Other build: \code{\link{module_build}},
  \code{\link{module_identities}},
  \code{\link{module_skeleton}}, \code{\link{module_test}},
  \code{\link{module_travis}}, \code{\link{module_upload}}
}
\concept{build}
