% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_identities}
\alias{module_identities}
\title{Return identities for a module}
\usage{
module_identities(flpth = getwd())
}
\arguments{
\item{flpth}{File path to location of module}
}
\value{
Logical
}
\description{
Returns a list of the identities (GitHub repo, Package name,
Docker images) for an outsider module. Works for modules in development.
Requires module to have a file path.
}
\examples{
library(outsider)

# build file structure for an example module
module_path <- module_skeleton(program_name = "goldenhind",
                               repo_user = "drake_on_github",
                               docker_user = "drake_on_docker",
                               full_name = 'Sir Francis Drake',
                               email = 'f.drake@goldenhind.gov.uk',
                               service = 'github',
                               flpth = tempdir())
# new path created 
(module_path)
# check the generated names and links
module_identities(flpth = module_path)
# check the files are in the right locations
module_check(flpth = module_path)
# deliberately break: delete a folder and check again
unlink(x = file.path(module_path, 'inst'), recursive = TRUE, force = TRUE)
module_check(flpth = module_path)

# clean-up
unlink(x = module_path, recursive = TRUE, force = TRUE)
}
\seealso{
Other build: 
\code{\link{module_build}()},
\code{\link{module_check}()},
\code{\link{module_skeleton}()},
\code{\link{module_test}()},
\code{\link{module_travis}()},
\code{\link{module_upload}()}
}
\concept{build}
