% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_travis}
\alias{module_travis}
\title{Generate Travis-CI file (GitHub only)}
\usage{
module_travis(flpth = getwd())
}
\arguments{
\item{flpth}{Directory in which to create .travis.yml}
}
\value{
Logical
}
\description{
Write .travis.yml to working directory.
}
\details{
Validated outsider modules must have a .travis.yml in their
repository. These .travis.yml are created with \link{module_skeleton} but
can also be generated using this function.
}
\seealso{
Other build: \code{\link{module_build}},
  \code{\link{module_check}},
  \code{\link{module_identities}},
  \code{\link{module_skeleton}}, \code{\link{module_test}},
  \code{\link{module_upload}}
}
\concept{build}
