% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_upload}
\alias{module_upload}
\title{Upload a module to code sharing site and DockerHub}
\usage{
module_upload(
  flpth = getwd(),
  code_sharing = TRUE,
  dockerhub = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{flpth}{File path to location of module}

\item{code_sharing}{Upload to code sharing service?}

\item{dockerhub}{Upload to DockerHub?}

\item{verbose}{Print docker and program info to console}
}
\value{
Logical
}
\description{
Look up usernames and other information contained in
"om.yml" to upload module to a code sharing site (github, gitlab or
bitbucket) and/or DockerHub.
}
\details{
This function runs \code{\link{git_upload}} and
\code{\link{docker_push}}.
}
\examples{
library(outsider)

# NOT RUN
# # build a simple module
# module_path <- module_skeleton(program_name = 'echo', flpth = getwd(),
#                                repo_user = '[YOUR USERNAME]',
#                                docker_user = '[YOUR USERNAME]',
#                                service = 'github')
# module_build(module_path)
# # upload to your GitHub and your DockerHub
# module_upload(flpth = module_path)
}
\seealso{
Other build: 
\code{\link{module_build}()},
\code{\link{module_check}()},
\code{\link{module_identities}()},
\code{\link{module_skeleton}()},
\code{\link{module_test}()},
\code{\link{module_travis}()}
}
\concept{build}
