% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.devtools.R
\docType{package}
\name{outsider.devtools}
\alias{outsider.devtools}
\alias{outsider.devtools-package}
\title{outsider.devtools: Build 'outsider' Modules}
\description{
Tools, resources and information for making it easier to build your own
outsider modules.
}
\details{
For more information visit the outsider website
(\url{https://antonellilab.github.io/outsider.devtools/}).
}
