% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgdetails_get}
\alias{pkgdetails_get}
\title{Read the package description}
\usage{
pkgdetails_get(flpth)
}
\arguments{
\item{flpth}{Path to package}
}
\value{
list
}
\description{
Return a list of all package details based on a package's
DESCRIPTION file.
}
\seealso{
Other utils: \code{\link{description_get}},
  \code{\link{examples_get}}, \code{\link{file_create}},
  \code{\link{pkgnm_get}}, \code{\link{string_replace}},
  \code{\link{tags_get}}, \code{\link{templates_get}},
  \code{\link{yaml_get}}
}
\concept{utils}
