% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgnm_get}
\alias{pkgnm_get}
\title{Determine R package name}
\usage{
pkgnm_get(flpth)
}
\arguments{
\item{flpth}{Path to module}
}
\value{
character(1)
}
\description{
Determine R package name from file path
}
\seealso{
Other utils: \code{\link{description_get}},
  \code{\link{examples_get}}, \code{\link{file_create}},
  \code{\link{pkgdetails_get}},
  \code{\link{string_replace}}, \code{\link{tags_get}},
  \code{\link{templates_get}}, \code{\link{yaml_get}}
}
\concept{utils}
