% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_base.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arglist_get}
\alias{arglist_parse}
\alias{dirpath_get}
\alias{filestosend_get}
\alias{wd_get}
\alias{docker_img_ls}
\alias{image_install}
\alias{is_docker_available}
\alias{outsider_init}
\alias{run}
\alias{is_installed}
\alias{meta_get}
\alias{modules_list}
\alias{pkg_install}
\alias{uninstall}
\alias{cat_line}
\alias{char}
\alias{func}
\alias{log_set}
\alias{stat}
\alias{server_connect}
\alias{server_disconnect}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{outsider.base}{\code{\link[outsider.base]{arglist_get}}, \code{\link[outsider.base]{arglist_parse}}, \code{\link[outsider.base]{cat_line}}, \code{\link[outsider.base]{char}}, \code{\link[outsider.base]{dirpath_get}}, \code{\link[outsider.base]{docker_img_ls}}, \code{\link[outsider.base]{filestosend_get}}, \code{\link[outsider.base]{func}}, \code{\link[outsider.base]{image_install}}, \code{\link[outsider.base]{is_docker_available}}, \code{\link[outsider.base]{is_installed}}, \code{\link[outsider.base]{log_set}}, \code{\link[outsider.base]{meta_get}}, \code{\link[outsider.base]{modules_list}}, \code{\link[outsider.base]{outsider_init}}, \code{\link[outsider.base]{pkg_install}}, \code{\link[outsider.base]{run}}, \code{\link[outsider.base]{server_connect}}, \code{\link[outsider.base]{server_disconnect}}, \code{\link[outsider.base]{stat}}, \code{\link[outsider.base]{uninstall}}, \code{\link[outsider.base]{wd_get}}}
}}

