% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{remote_git_exists}
\alias{remote_git_exists}
\title{Does a remote git repo exist}
\usage{
remote_git_exists(url)
}
\arguments{
\item{url}{URL to check}
}
\value{
Logical
}
\description{
Return TRUE if request returns a 200 status code.
}
\details{
Private repositories will not be discovered. Doesn't work for
bitbucket.
}
\seealso{
Other git: 
\code{\link{git_upload}()}
}
\concept{git}
