% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_replace}
\alias{string_replace}
\title{Replace patterns in a string}
\usage{
string_replace(string, patterns, values)
}
\arguments{
\item{string}{Text}

\item{patterns}{Patterns to replace with values}

\item{values}{Values to be put in place}
}
\value{
character
}
\description{
For a given character string, replace patterns with values.
}
\seealso{
Other utils: \code{\link{description_get}},
  \code{\link{examples_get}}, \code{\link{file_create}},
  \code{\link{pkgdetails_get}}, \code{\link{pkgnm_get}},
  \code{\link{tags_get}}, \code{\link{templates_get}},
  \code{\link{yaml_get}}
}
\concept{utils}
