% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tags_get}
\alias{tags_get}
\title{Determine tags}
\usage{
tags_get(flpth)
}
\arguments{
\item{flpth}{Path to module}
}
\value{
character
}
\description{
Determine module Docker tags
}
\seealso{
Other utils: 
\code{\link{description_get}()},
\code{\link{examples_get}()},
\code{\link{file_create}()},
\code{\link{pkgdetails_get}()},
\code{\link{pkgnm_get}()},
\code{\link{string_replace}()},
\code{\link{templates_get}()},
\code{\link{yaml_get}()}
}
\concept{utils}
