% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{templates_get}
\alias{templates_get}
\title{Retrieve template files}
\usage{
templates_get()
}
\value{
character vector
}
\description{
Return template files for an outsider module.
}
\seealso{
Other utils: 
\code{\link{description_get}()},
\code{\link{examples_get}()},
\code{\link{file_create}()},
\code{\link{pkgdetails_get}()},
\code{\link{pkgnm_get}()},
\code{\link{string_replace}()},
\code{\link{tags_get}()},
\code{\link{yaml_get}()}
}
\concept{utils}
