% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yaml_get}
\alias{yaml_get}
\title{Return om.yml}
\usage{
yaml_get(flpth)
}
\arguments{
\item{flpth}{Path to module}
}
\value{
list
}
\description{
Return contents of om.yml
}
\seealso{
Other utils: \code{\link{description_get}},
  \code{\link{examples_get}}, \code{\link{file_create}},
  \code{\link{pkgdetails_get}}, \code{\link{pkgnm_get}},
  \code{\link{string_replace}}, \code{\link{tags_get}},
  \code{\link{templates_get}}
}
\concept{utils}
