% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_build}
\alias{docker_build}
\title{Build a Docker image}
\usage{
docker_build(img, tag, url_or_path, verbose)
}
\arguments{
\item{img}{Image name}

\item{tag}{Tag, e.g. 'latest'}

\item{url_or_path}{URL or file path to Dockerfile}

\item{verbose}{Be verbose? T/F}
}
\value{
Logical
}
\description{
Build a Docker image through system call.
}
\seealso{
Other docker: 
\code{\link{docker_push}()}
}
\concept{docker}
