% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_push}
\alias{docker_push}
\title{Push a Docker image to Docker Hub}
\usage{
docker_push(username, img, tag, verbose)
}
\arguments{
\item{username}{Login username for Docker Hub.}

\item{img}{Image name}

\item{tag}{Tag, e.g. 'latest'}

\item{verbose}{Be verbose? T/F}
}
\value{
Logical
}
\description{
Push a Docker image to Docker Hub. Requires a user to have login
details for Docker Hub, \url{https://hub.docker.com/}.
Returns TRUE if no errors are raised.
}
\examples{
library(outsider.devtools)

# NOT RUN
# # create image from om..echo Docker file
# url <- paste0('https://raw.githubusercontent.com/DomBennett/om..hello.world/',
#               'master/inst/dockerfiles/latest/Dockerfile')
# docker_build(img = 'test_docker_push', tag = 'latest', verbose = TRUE,
#              url_or_path = url)
# # push to your Docker Hub
# docker_push(username = '[YOUR USERNAME]', img = 'test_docker_push',
#             tag = 'latest', verbose = TRUE)
}
\seealso{
Other docker: 
\code{\link{docker_build}()}
}
\concept{docker}
