% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_upload}
\alias{git_upload}
\title{Upload module to code-sharing service}
\usage{
git_upload(flpth, username, service = c("github", "gitlab", "bitbucket"))
}
\arguments{
\item{flpth}{File path to module.}

\item{username}{Username for code-sharing service.}

\item{service}{Code-sharing service}
}
\value{
Logical
}
\description{
Upload module to a git-based code-sharing service. Initiate
a git repo, add core module files, commit and push to remote.
}
\details{
Remote URL is determined to be: code sharing URL + username + R
package name.
}
\seealso{
Other git: \code{\link{remote_git_exists}}
}
\concept{git}
