% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_build}
\alias{module_build}
\title{Build a module}
\usage{
module_build(
  flpth = getwd(),
  tag = "latest",
  build_documents = TRUE,
  build_package = TRUE,
  build_image = TRUE,
  build_readme = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{flpth}{File path to location of module.}

\item{tag}{Docker tag, e.g. latest.}

\item{build_documents}{Build R documentation? T/F}

\item{build_package}{Build R package? T/F}

\item{build_image}{Build Docker image? T/F}

\item{build_readme}{Build README.md? T/F}

\item{verbose}{Be verbose? T/F}
}
\value{
Logical
}
\description{
Do
}
\examples{
library(outsider)

# NOT RUN
# # build a skeleton package
# module_path <- module_skeleton(program_name = 'echo', flpth = getwd())
# # check the file structure
# module_check(flpth = module_path)
# # look-up key identifying names: R package name, Docker image name
# module_identities(flpth = module_path)
# # build the R package and Docker image
# module_build(flpth = module_path, tag = 'latest')
# # test the module
# module_test(flpth = module_path)
# # clean-up
# unlink(x = module_path, recursive = TRUE)
}
\seealso{
Other build: 
\code{\link{module_check}()},
\code{\link{module_identities}()},
\code{\link{module_skeleton}()},
\code{\link{module_test}()},
\code{\link{module_travis}()},
\code{\link{module_upload}()}
}
\concept{build}
