% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_build}
\alias{module_build}
\title{Build a module}
\usage{
module_build(flpth, tag = NULL, build_documents = TRUE,
  build_package = TRUE, build_image = TRUE, verbose = TRUE)
}
\arguments{
\item{flpth}{File path to location of module.}

\item{tag}{Docker tag, e.g. latest.}

\item{build_documents}{Build R documentation? T/F}

\item{build_package}{Build R package? T/F}

\item{build_image}{Build Docker image? T/F}

\item{verbose}{Be verbose? T/F}
}
\value{
Logical
}
\description{
Do
}
\examples{
library(outsider)
# build a skeleton package
flpth <- tempdir()
mdl_pth <- module_skeleton(program_name = 'example', repo_user = 'github_user',
                           docker_user = 'docker_user', flpth = flpth)
# look-up module user and program name/ids
# note: "om.." is added to the beginning of the program name
# mdl_pth <- file.path(flpth, 'om..example')
# (ids <- module_identities(flpth = mdl_pth))
# # generate .travis.yml
# module_travis(repo = ids$`GitHub repo`, flpth = mdl_pth)
# cat(readLines(file.path(mdl_pth, '.travis.yml')), sep = '\\n')
# # check the package is formatted correctly
# module_check(flpth = mdl_pth)


# clean-up
unlink(x = mdl_pth, recursive = TRUE)
}
\seealso{
Other build: \code{\link{module_check}},
  \code{\link{module_identities}},
  \code{\link{module_skeleton}}, \code{\link{module_test}},
  \code{\link{module_travis}}, \code{\link{module_upload}}
}
\concept{build}
