% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_check}
\alias{module_check}
\title{Check names and structure of a module}
\usage{
module_check(flpth = getwd())
}
\arguments{
\item{flpth}{File path to location of module}
}
\value{
Logical
}
\description{
Returns TRUE if all the names and structure of a minimal viable
outsider module are correct.
}
\examples{
library(outsider)

# NOT RUN
# # build a skeleton package
# module_path <- module_skeleton(program_name = 'echo', flpth = getwd())
# # check the file structure
# module_check(flpth = module_path)
# # look-up key identifying names: R package name, Docker image name
# module_identities(flpth = module_path)
# # build the R package and Docker image
# module_build(flpth = module_path, tag = 'latest')
# # test the module
# module_test(flpth = module_path)


# clean-up
unlink(x = module_path, recursive = TRUE)
}
\seealso{
Other build: 
\code{\link{module_build}()},
\code{\link{module_identities}()},
\code{\link{module_skeleton}()},
\code{\link{module_test}()},
\code{\link{module_travis}()},
\code{\link{module_upload}()}
}
\concept{build}
