% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_travis}
\alias{module_travis}
\title{Generate Travis-CI file}
\usage{
module_travis(repo, flpth = getwd())
}
\arguments{
\item{repo}{Repository}

\item{flpth}{Directory in which to create .travis.yml}
}
\value{
Logical
}
\description{
Write .travis.yml to working directory.
}
\details{
All validated outsider modules must have a .travis.yml in their
repository. These .travis.yml must be generated using this function.
}
