% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_travis}
\alias{module_travis}
\title{Generate Travis-CI file (GitHub only)}
\usage{
module_travis(flpth = getwd())
}
\arguments{
\item{flpth}{Directory in which to create .travis.yml}
}
\value{
Logical
}
\description{
Write .travis.yml to working directory.
}
\details{
Validated outsider modules must have a .travis.yml in their
repository. These \code{.travis.yml} are created with \link{module_skeleton}
but can also be generated using this function.
}
\examples{
library(outsider)

# a skeleton already comes with a .travis.yml
module_path <- module_skeleton(program_name = "goldenhind",
                               repo_user = "drake_on_github",
                               docker_user = "drake_on_docker",
                               service = 'github',
                               flpth = tempdir())
(file.exists(file.path(module_path, '.travis.yml')))
# but if it were deleted, needs updating or mistakenly edited ...
file.remove(file.path(module_path, '.travis.yml'))
# a new one can generated
module_travis(flpth = module_path)
}
\seealso{
Other build: 
\code{\link{module_build}()},
\code{\link{module_check}()},
\code{\link{module_identities}()},
\code{\link{module_skeleton}()},
\code{\link{module_test}()},
\code{\link{module_upload}()}
}
\concept{build}
