% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{templates_get}
\alias{templates_get}
\title{Retrieve template files}
\usage{
templates_get()
}
\value{
character vector
}
\description{
Return template files for an outsider module.
}
\seealso{
Other private: \code{\link{examples_test}},
  \code{\link{file_create}}, \code{\link{pkgdetails_get}},
  \code{\link{string_replace}}, \code{\link{test}}
}
\concept{private}
