% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test a module}
\usage{
test(flpth)
}
\arguments{
\item{repo}{Repository}
}
\value{
logical
}
\description{
Test an outsider module by making sure it installs,
imports and its examples run correctly.
}
\seealso{
Other private: \code{\link{examples_test}},
  \code{\link{file_create}}, \code{\link{fnames_get}},
  \code{\link{pkgdetails_get}},
  \code{\link{string_replace}}, \code{\link{templates_get}}
}
\concept{private}
