% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test a module}
\usage{
test(flpth, pull = FALSE)
}
\arguments{
\item{flpth}{File path to location of module}

\item{pull}{Pull Docker image from Docker Hub? T/F}
}
\value{
logical
}
\description{
Test an outsider module by making sure it installs,
imports and its examples run correctly. Raises an error if a test fails.
}
\seealso{
Other testing: 
\code{\link{examples_test}()}
}
\concept{testing}
